# frontend/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "HuntTrackUpdateRequest": {
        "properties": {
          "acquisition_status": {
            "title": "Acquisition Status",
            "type": "string"
          }
        },
        "required": [
          "acquisition_status"
        ],
        "title": "HuntTrackUpdateRequest",
        "type": "object"
      },
...
      "TrackAffinitiesResponse": {
        "properties": {
          "affinities": {
            "items": {
              "$ref": "#/components/schemas/AffinityListItem"
            },
            "title": "Affinities",
            "type": "array"
          }
        },
        "required": [
          "affinities"
        ],
        "title": "TrackAffinitiesResponse",
        "type": "object"
      },
      "TrackAffinityRequest": {
        "properties": {
          "affinity": {
            "title": "Affinity",
            "type": "string"
          },
          "other_track_id": {
            "title": "Other Track Id",
            "type": "integer"
          }
        },
        "required": [
          "other_track_id",
          "affinity"
        ],
        "title": "TrackAffinityRequest",
        "type": "object"
      },
      "TrackAffinityResponse": {
        "properties": {
          "affinity": {
            "title": "Affinity",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "track_a_id": {
            "title": "Track A Id",
            "type": "integer"
          },
          "track_b_id": {
            "title": "Track B Id",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "track_a_id",
          "track_b_id",
          "affinity"
        ],
        "title": "TrackAffinityResponse",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Kiku API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/hunt/tracks/{track_id}": {
      "patch": {
        "description": "Update a hunt track's acquisition status (e.g. mark as 'wanted').",
        "operationId": "update_hunt_track_api_hunt_tracks__track_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "track_id",
            "required": true,
            "schema": {
              "title": "Track Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/HuntTrackUpdateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HuntTrackResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Hunt Track",
        "tags": [
          "hunt"
        ]
      }
    },
...
    "/api/sets/{set_id}/tracks/{track_id}": {
      "delete": {
        "description": "Remove a track from a set, returning the set's tracks as they now stand.\n\nLike add and reorder, this hands back the whole new list: the removal shifts\nevery later position and rescores the transition that closed over the gap, so\nthe caller would otherwise have to re-read the set to learn what it now is.",
        "operationId": "remove_track_api_sets__set_id__tracks__track_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "set_id",
            "required": true,
            "schema": {
              "title": "Set Id",
              "type": "integer"
            }
          },
          {
            "in": "path",
            "name": "track_id",
            "required": true,
            "schema": {
              "title": "Track Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/SetTrackResponse"
                  },
                  "title": "Response Remove Track Api Sets  Set Id  Tracks  Track Id  Delete",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Remove Track",
        "tags": [
          "sets"
        ]
      }
    },
...
    "/api/tracks/autocomplete/artists": {
      "get": {
        "operationId": "artists_autocomplete_api_tracks_autocomplete_artists_get",
        "parameters": [
          {
            "in": "query",
            "name": "q",
            "required": true,
            "schema": {
              "minLength": 1,
              "title": "Q",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 50,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "type": "string"
                  },
                  "title": "Response Artists Autocomplete Api Tracks Autocomplete Artists Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Artists Autocomplete",
        "tags": [
          "tracks"
        ]
      }
    },
    "/api/tracks/autocomplete/labels": {
      "get": {
        "operationId": "labels_autocomplete_api_tracks_autocomplete_labels_get",
        "parameters": [
          {
            "in": "query",
            "name": "q",
            "required": true,
            "schema": {
              "minLength": 1,
              "title": "Q",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 50,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "type": "string"
                  },
                  "title": "Response Labels Autocomplete Api Tracks Autocomplete Labels Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Labels Autocomplete",
        "tags": [
          "tracks"
        ]
      }
    },
...
    "/api/tracks/{track_id}": {
      "get": {
        "operationId": "track_detail_api_tracks__track_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "track_id",
            "required": true,
            "schema": {
              "title": "Track Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TrackResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Track Detail",
        "tags": [
          "tracks"
        ]
      }
    },
    "/api/tracks/{track_id}/affinities": {
      "get": {
        "description": "Return all affinities for a track, with full TrackResponse for each partner.",
        "operationId": "list_affinities_api_tracks__track_id__affinities_get",
        "parameters": [
          {
            "in": "path",
            "name": "track_id",
            "required": true,
            "schema": {
              "title": "Track Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TrackAffinitiesResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Affinities",
        "tags": [
          "tracks"
        ]
      }
    },
    "/api/tracks/{track_id}/affinity": {
      "post": {
        "description": "Mark two tracks as 'good' or 'bad' together. Upserts if pair exists.",
        "operationId": "set_affinity_api_tracks__track_id__affinity_post",
        "parameters": [
          {
            "in": "path",
            "name": "track_id",
            "required": true,
            "schema": {
              "title": "Track Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TrackAffinityRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TrackAffinityResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Set Affinity",
        "tags": [
          "tracks"
        ]
      }
    },
    "/api/tracks/{track_id}/affinity/{other_id}": {
      "delete": {
        "description": "Remove affinity between two tracks.",
        "operationId": "delete_affinity_api_tracks__track_id__affinity__other_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "track_id",
            "required": true,
            "schema": {
              "title": "Track Id",
              "type": "integer"
            }
          },
          {
            "in": "path",
            "name": "other_id",
            "required": true,
            "schema": {
              "title": "Other Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Affinity",
        "tags": [
          "tracks"
        ]
      }
    },
    "/api/tracks/{track_id}/artwork": {
      "get": {
        "description": "Serve a track's artwork: its own embedded art, else its album's cover.\n\nNever 500s \u2014 embedded extraction soft-fails to None, and a track with no\nembedded art inherits the album cover via the resolver (so the Tracks table\nand Now Playing bar fill in too). A clean 404 is the only failure the client\nsees; the frontend hides the <img> on 404.",
        "operationId": "track_artwork_api_tracks__track_id__artwork_get",
        "parameters": [
          {
            "in": "path",
            "name": "track_id",
            "required": true,
            "schema": {
              "title": "Track Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Track Artwork",
        "tags": [
          "tracks"
        ]
      }
    },
...
  }
}